requirements.txt
aiohttp == 3.6.2
Pillow == 7.1.2
requests == 2.23.0
